Refactored common code into BaseTest
